+++ Runtime/RendererFeatures/DecalRendererFeature.cs
@@ -1,3 +1,4 @@
+using System.Diagnostics;
 using UnityEngine.Assertions;
 using UnityEngine.Rendering.Universal.Internal;
 
@@ -310,24 +311,15 @@ namespace UnityEngine.Rendering.Universal
 
         private bool IsAutomaticDBuffer()
         {
+            // As WebGL uses gles here we should not use DBuffer
 #if UNITY_EDITOR
-            var selectedBuildTargetGroup = UnityEditor.EditorUserBuildSettings.selectedBuildTargetGroup;
-            if (selectedBuildTargetGroup == UnityEditor.BuildTargetGroup.Standalone)
-                return true;
-            if (selectedBuildTargetGroup == UnityEditor.BuildTargetGroup.GameCoreXboxOne)
-                return true;
-            if (selectedBuildTargetGroup == UnityEditor.BuildTargetGroup.GameCoreXboxSeries)
-                return true;
-            if (selectedBuildTargetGroup == UnityEditor.BuildTargetGroup.PS4)
-                return true;
-            if (selectedBuildTargetGroup == UnityEditor.BuildTargetGroup.PS5)
-                return true;
-            if (selectedBuildTargetGroup == UnityEditor.BuildTargetGroup.WSA)
-                return true;
-            return false;
+            if (UnityEditor.EditorUserBuildSettings.selectedBuildTargetGroup == UnityEditor.BuildTargetGroup.WebGL)
+                return false;
 #else
-            return SystemInfo.deviceType == DeviceType.Desktop || SystemInfo.deviceType == DeviceType.Console;
+            if (Application.platform == RuntimePlatform.WebGLPlayer)
+                return false;
 #endif
+            return !GraphicsSettings.HasShaderDefine(BuiltinShaderDefine.SHADER_API_MOBILE);
         }
 
         private void RecreateSystemsIfNeeded(ScriptableRenderer renderer, in CameraData cameraData)
@@ -392,7 +384,10 @@ namespace UnityEngine.Rendering.Universal
                     m_ForwardEmissivePass = new DecalForwardEmissivePass(m_DecalDrawForwardEmissiveSystem);
 
                     if (universalRenderer.actualRenderingMode == RenderingMode.Deferred)
+                    {
                         m_DBufferRenderPass.deferredLights = universalRenderer.deferredLights;
+                        m_DBufferRenderPass.deferredLights.DisableFramebufferFetchInput();
+                    }
                     break;
             }
 
@@ -405,6 +400,8 @@ namespace UnityEngine.Rendering.Universal
                 return;
 
             RecreateSystemsIfNeeded(renderer, cameraData);
+
+            ChangeAdaptivePerformanceDrawDistances();
 
             m_DecalEntityManager.Update();
 
@@ -442,6 +439,8 @@ namespace UnityEngine.Rendering.Universal
             }
 
             RecreateSystemsIfNeeded(renderer, renderingData.cameraData);
+
+            ChangeAdaptivePerformanceDrawDistances();
 
             if (intermediateRendering)
             {
@@ -483,6 +482,10 @@ namespace UnityEngine.Rendering.Universal
             }
         }
 
+        internal override bool SupportsNativeRenderPass()
+        {
+            return m_Technique == DecalTechnique.GBuffer || m_Technique == DecalTechnique.ScreenSpace;
+        }
         protected override void Dispose(bool disposing)
         {
             CoreUtils.Destroy(m_CopyDepthMaterial);
@@ -493,6 +496,24 @@ namespace UnityEngine.Rendering.Universal
                 m_DecalEntityManager = null;
                 sharedDecalEntityManager.Release(m_DecalEntityManager);
             }
+        }
+
+        [Conditional("ADAPTIVE_PERFORMANCE_4_0_0_OR_NEWER")]
+        private void ChangeAdaptivePerformanceDrawDistances()
+        {
+#if ADAPTIVE_PERFORMANCE_4_0_0_OR_NEWER
+            if (UniversalRenderPipeline.asset.useAdaptivePerformance)
+            {
+                if (m_DecalCreateDrawCallSystem != null)
+                {
+                    m_DecalCreateDrawCallSystem.maxDrawDistance = AdaptivePerformance.AdaptivePerformanceRenderSettings.DecalsDrawDistance;
+                }
+                if (m_DecalUpdateCullingGroupSystem != null)
+                {
+                    m_DecalUpdateCullingGroupSystem.boundingDistance = AdaptivePerformance.AdaptivePerformanceRenderSettings.DecalsDrawDistance;
+                }
+            }
+#endif
         }
     }
 }