+++ Runtime/UniversalRenderPipelineGlobalSettings.cs
@@ -115,6 +115,7 @@ namespace UnityEngine.Rendering.Universal
                         AssetDatabase.CreateFolder("Assets", folderPath);
                     assetCreated = Create(path);
 
+                    // TODO: Reenable after next urp template is published
                     //Debug.LogWarning("No URP Global Settings Asset is assigned. One will be created for you. If you want to modify it, go to Project Settings > Graphics > URP Settings.");
                 }
                 else